Move regridding_utils to utilities/regrid
