+++ imod/mf6/drn.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from imod.mf6.boundary_condition import BoundaryCondition
-from imod.mf6.regridding_utils import RegridderType
+from imod.mf6.utilities.regrid import RegridderType
 from imod.mf6.validation import BOUNDARY_DIMS_SCHEMA, CONC_DIMS_SCHEMA
 from imod.schemata import (
     AllInsideNoDataSchema,
