+++ imod/mf6/evt.py
@@ -3,7 +3,7 @@ from typing import Optional
 import numpy as np
 
 from imod.mf6.boundary_condition import BoundaryCondition
-from imod.mf6.regridding_utils import RegridderType
+from imod.mf6.utilities.regrid import RegridderType
 from imod.mf6.validation import BOUNDARY_DIMS_SCHEMA, CONC_DIMS_SCHEMA
 from imod.schemata import (
     AllInsideNoDataSchema,
