+++ imod/mf6/npf.py
@@ -3,7 +3,7 @@ import warnings
 import numpy as np
 
 from imod.mf6.package import Package
-from imod.mf6.regridding_utils import RegridderType
+from imod.mf6.utilities.regrid import RegridderType
 from imod.mf6.validation import PKG_DIMS_SCHEMA
 from imod.schemata import (
     AllValueSchema,
