+++ imod/mf6/oc.py
@@ -6,8 +6,8 @@ from typing import Tuple, Union
 import numpy as np
 
 from imod.mf6.package import Package
-from imod.mf6.regridding_utils import RegridderType
 from imod.mf6.utilities.dataset import is_dataarray_none
+from imod.mf6.utilities.regrid import RegridderType
 from imod.mf6.write_context import WriteContext
 from imod.schemata import DTypeSchema
